Split hbit and herc20 db modules
